refactor:page container moved to page
