+++ packages/widget/src/components/SendToWallet/SendToWallet.styled.tsx
@@ -1,8 +1,4 @@
-import {
-  Container,
-  InputBase,
-  FormControl as MuiFormControl,
-} from '@mui/material';
+import { InputBase, FormControl as MuiFormControl } from '@mui/material';
 import Alert from '@mui/material/Alert';
 import { inputBaseClasses } from '@mui/material/InputBase';
 import { styled } from '@mui/material/styles';
@@ -22,12 +18,6 @@ export const Input = styled(InputBase)(({ theme }) => ({
   [`.${inputBaseClasses.input}.${inputBaseClasses.disabled}`]: {
     WebkitTextFillColor: 'unset',
   },
-}));
-
-export const PageContainer = styled(Container)(({ theme }) => ({
-  height: 508,
-  display: 'flex',
-  flexDirection: 'column',
 }));
 
 export const AlertSection = styled(Alert)(({ theme }) => ({
