+++ packages/widget/src/pages/SendToWalletPage/SendToWalletPage.styled.tsx
@@ -1,0 +1,8 @@
+import styled from '@emotion/styled';
+import { Container } from '@mui/material';
+
+export const PageContainer = styled(Container)(({ theme }) => ({
+  height: 508,
+  display: 'flex',
+  flexDirection: 'column',
+}));
