+++ packages/widget/src/pages/SendToWalletPage/SendToWalletPage.tsx
@@ -1,5 +1,10 @@
 import { SendToWallet } from '../../components/SendToWallet';
+import { PageContainer } from './SendToWalletPage.styled';
 
 export const SendToWalletPage = () => {
-  return <SendToWallet />;
+  return (
+    <PageContainer>
+      <SendToWallet />
+    </PageContainer>
+  );
 };
